+++ samples/Sentry.Samples.AspNetCore.Basic/Program.cs
@@ -15,12 +15,14 @@ builder.WebHost.UseSentry(options =>
     // Log debug information about the Sentry SDK
     options.Debug = true;
 #endif
+
+    // This option enables Logs sent to Sentry.
+    options.Experimental.EnableLogs = true;
 });
 
 var app = builder.Build();
 
-// An example ASP.NET Core middleware that throws an
-// exception when serving a request to path: /throw
+// An example ASP.NET Core endpoint that throws an exception when serving a request to path: /throw
 app.MapGet("/throw/{message?}", context =>
 {
     var exceptionMessage = context.GetRouteValue("message") as string;
@@ -46,6 +48,27 @@ app.MapGet("/throw/{message?}", context =>
     // as exemplified above.
     throw new Exception(
         exceptionMessage ?? "An exception thrown from the ASP.NET Core pipeline");
+});
+
+// Demonstrates how to add tracing in custom middleware
+app.Use(async (context, next) =>
+{
+    var span = SentrySdk.StartSpan("CustomMiddlewareSpan", "middleware");
+    try
+    {
+        var log = context.RequestServices.GetRequiredService<ILoggerFactory>()
+            .CreateLogger<Program>();
+
+        log.LogInformation("Just chilling for a bit...");
+        await Task.Delay(TimeSpan.FromMilliseconds(50)); // Simulate some work
+        span.Finish();
+    }
+    catch (Exception e)
+    {
+        span.Finish(e);
+        throw;
+    }
+    await next();
 });
 
 app.Run();