+++ src/Sentry/Internal/Hub.cs
@@ -1,5 +1,7 @@
 using Sentry.Extensibility;
 using Sentry.Infrastructure;
+using Sentry.Integrations;
+using Sentry.Internal.Extensions;
 using Sentry.Protocol.Envelopes;
 using Sentry.Protocol.Metrics;
 
@@ -14,6 +16,7 @@ internal class Hub : IHub, IDisposable
     private readonly SentryOptions _options;
     private readonly RandomValuesFactory _randomValuesFactory;
     private readonly IReplaySession _replaySession;
+    private readonly List<IDisposable> _integrationsToCleanup = new();
 
 #if MEMORY_DUMP_SUPPORTED
     private readonly MemoryMonitor? _memoryMonitor;
@@ -66,6 +69,8 @@ internal class Hub : IHub, IDisposable
             PushScope();
         }
 
+        Logger = SentryStructuredLogger.Create(this, options, _clock);
+
 #if MEMORY_DUMP_SUPPORTED
         if (options.HeapDumpOptions is not null)
         {
@@ -84,6 +89,10 @@ internal class Hub : IHub, IDisposable
         {
             options.LogDebug("Registering integration: '{0}'.", integration.GetType().Name);
             integration.Register(this, options);
+            if (integration is IDisposable disposableIntegration)
+            {
+                _integrationsToCleanup.Add(disposableIntegration);
+            }
         }
     }
 
@@ -166,8 +175,8 @@ internal class Hub : IHub, IDisposable
 
         bool? isSampled = null;
         double? sampleRate = null;
-        var sampleRand = dynamicSamplingContext?.Items.TryGetValue("sample_rand", out var dscsampleRand) ?? false
-            ? double.Parse(dscsampleRand, NumberStyles.Float, CultureInfo.InvariantCulture)
+        var sampleRand = dynamicSamplingContext?.Items.TryGetValue("sample_rand", out var dscSampleRand) ?? false
+            ? double.Parse(dscSampleRand, NumberStyles.Float, CultureInfo.InvariantCulture)
             : SampleRandHelper.GenerateSampleRand(context.TraceId.ToString());
 
         // TracesSampler runs regardless of whether a decision has already been made, as it can be used to override it.
@@ -181,14 +190,26 @@ internal class Hub : IHub, IDisposable
             {
                 // The TracesSampler trumps all other sampling decisions (even the trace header)
                 sampleRate = samplerSampleRate;
-                isSampled = SampleRandHelper.IsSampled(sampleRand, sampleRate.Value);
+                isSampled = SampleRandHelper.IsSampled(sampleRand, samplerSampleRate);
+
+                // Ensure the actual sampleRate is set on the provided DSC (if any) when the TracesSampler reached a sampling decision
+                dynamicSamplingContext = dynamicSamplingContext?.WithSampleRate(samplerSampleRate);
             }
         }
 
         // If the sampling decision isn't made by a trace sampler we check the trace header first (from the context) or
         // finally fallback to Random sampling if the decision has been made by no other means
-        sampleRate ??= _options.TracesSampleRate ?? 0.0;
-        isSampled ??= context.IsSampled ?? SampleRandHelper.IsSampled(sampleRand, sampleRate.Value);
+        if (isSampled == null)
+        {
+            sampleRate = _options.TracesSampleRate ?? 0.0;
+            isSampled = context.IsSampled ?? SampleRandHelper.IsSampled(sampleRand, sampleRate.Value);
+
+            if (context.IsSampled is null && _options.TracesSampleRate is not null)
+            {
+                // Ensure the actual sampleRate is set on the provided DSC (if any) when not IsSampled upstream but the TracesSampleRate reached a sampling decision
+                dynamicSamplingContext = dynamicSamplingContext?.WithSampleRate(_options.TracesSampleRate.Value);
+            }
+        }
 
         // Make sure there is a replayId (if available) on the provided DSC (if any).
         dynamicSamplingContext = dynamicSamplingContext?.WithReplayId(_replaySession);
@@ -751,10 +772,33 @@ internal class Hub : IHub, IDisposable
         return SentryId.Empty;
     }
 
+    // Internal capture method that allows the Unity SDK to send attachments after an already captured event.
+    // Kept internal as the preferred way of adding attachments is either on the scope or directly on the event.
+    // See https://develop.sentry.dev/sdk/data-model/envelope-items/#attachment
+    internal bool CaptureAttachment(SentryId eventId, SentryAttachment attachment)
+    {
+        if (!IsEnabled || eventId == SentryId.Empty || attachment.IsNull())
+        {
+            return false;
+        }
+
+        try
+        {
+            var envelope = Envelope.FromAttachment(eventId, attachment, _options.DiagnosticLogger);
+            return CaptureEnvelope(envelope);
+        }
+        catch (Exception e)
+        {
+            _options.LogError(e, "Failure to capture attachment");
+            return false;
+        }
+    }
+
     public async Task FlushAsync(TimeSpan timeout)
     {
         try
         {
+            Logger.Flush();
             await CurrentClient.FlushAsync(timeout).ConfigureAwait(false);
         }
         catch (Exception e)
@@ -772,9 +816,24 @@ internal class Hub : IHub, IDisposable
             return;
         }
 
+        foreach (var integration in _integrationsToCleanup)
+        {
+            try
+            {
+                integration.Dispose();
+            }
+            catch (Exception e)
+            {
+                _options.LogError("Failed to dispose integration {0}: {1}", integration.GetType().Name, e);
+            }
+        }
+
 #if MEMORY_DUMP_SUPPORTED
         _memoryMonitor?.Dispose();
 #endif
+
+        Logger.Flush();
+        (Logger as IDisposable)?.Dispose(); // see Sentry.Internal.DefaultSentryStructuredLogger
 
         try
         {
@@ -800,4 +859,6 @@ internal class Hub : IHub, IDisposable
     }
 
     public SentryId LastEventId => CurrentScope.LastEventId;
+
+    public SentryStructuredLogger Logger { get; }
 }