+++ test/Sentry.Testing/NullAttachmentContent.cs
@@ -1,0 +1,10 @@
+namespace Sentry.Testing;
+
+internal sealed class NullAttachmentContent : IAttachmentContent
+{
+    public static NullAttachmentContent Instance { get; } = new();
+
+    public Stream GetStream() => Stream.Null;
+
+    private NullAttachmentContent() { }
+}